mainwindow refactor: - separate table handler class: handler UI and triggers API fetches - bkk api worker on a dedicated thread
