+++ meta-Qt/recipes-Qt/Bkk_Qt_App/files/src/bkk_worker_thread.hpp
@@ -7,7 +7,7 @@
 #include <atomic>
 #include <string>
 
-#include "bkk_api_wrapper.hpp"
+#include "bkk_api_worker.hpp"
 #include "bkk_clock_update.hpp"
 #include "bkk_online_check.hpp"
 
@@ -18,17 +18,13 @@ class WorkerThread : public QThread
 public:
     explicit WorkerThread(QObject *parent = nullptr);
 
-    void requestApiFetch();
     void requestClockUpdate();
     void requestOnlineCheck();
 
-    std::vector<StationArrival> getArrivals();
-    BkkApiError getErrorCode() const;
     std::string getClockText() const;
     bool isOnline() const;
 
 signals:
-    void apiFetchCompleted();
     void clockUpdateCompleted();
     void onlineCheckCompleted();
 
@@ -38,16 +34,12 @@ protected:
 private:
     mutable QMutex resultMutex;
 
-    BkkApiWrapper apiWrapper;
     CLockUpdater clockUpdater;
     OnlineChecker onlineChecker;
 
-    std::vector<StationArrival> arrivals;
-    BkkApiError errorCode;
     std::string clockText;
     bool onlineStatus;
 
-    std::atomic<bool> apiFetchRequested;
     std::atomic<bool> clockUpdateRequested;
     std::atomic<bool> onlineCheckRequested;
 };
